1.版本更新v1.0.0。

+++ bsp/wch/driver/board.c
@@ -6,14 +6,14 @@
  * @date 2024-01-16    MacRsh       First version
  */
 
-#include "mr_board.h"
+#include "../mr-library/driver/include/mr_board.h"
 
-void mr_delay_us(uint32_t us)
+void mr_delay_us(size_t us)
 {
     Delay_Us(us);
 }
 
-void mr_delay_ms(uint32_t ms)
+void mr_delay_ms(size_t ms)
 {
     Delay_Ms(ms);
 }
